Activate elfeed-ref-cache
An cache for the archive was introduced in commit 488e1966, which avoids
repeated reading the archive file when we open up entries one by one.
However, the cache is cleared as soon as idle time is observed, which is
sure to happen when we open one entry, take a glance, and open the next
entry.